pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "faiss-cpu>=1.10.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "flask-socketio>=5.5.1",
    "docx>=0.2.4",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "openai>=1.70.0",
    "eventlet>=0.39.1",
    "llama-cpp-python>=0.3.8",
    "tqdm>=4.67.1",
    "requests>=2.32.3",
]
